# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AQAR API Documentation",
        "description": "Multi-tenant real estate management API with three authentication guards: Admin, User (Landlord/Tenant Owner), and Tenant User.",
        "contact": {
            "name": "API Support",
            "email": "support@aqar.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.aqar.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/admins": {
            "get": {
                "tags": [
                    "Admin Management"
                ],
                "summary": "List all admins",
                "description": "Get paginated list of all administrators with optional filters",
                "operationId": "9d45e807b64639efce22090974258461",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name, email, or username",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "john"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filter by role name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "admin"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (active/inactive)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admins retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admins retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AdminResource"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum_admin": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Management"
                ],
                "summary": "Create new admin",
                "description": "Create a new administrator account",
                "operationId": "21808be8985ec0e98d4b16c901d8b18e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAdminRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AdminResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "This email is already registered."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum_admin": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreAdminRequest": {
                "title": "Store Admin Request",
                "required": [
                    "name",
                    "email",
                    "username",
                    "password",
                    "role"
                ],
                "properties": {
                    "name": {
                        "description": "Store Admin Request DTO\n\nValidates admin creation data.",
                        "type": "string",
                        "maxLength": 191,
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 191,
                        "example": "john@aqar.com"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "johndoe"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "SecurePass123!"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "SecurePass123!"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "mobile": {
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "example": "uploads/admins/avatar.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AdminResource": {
                "title": "Admin Resource",
                "description": "Admin user resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@aqar.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "admin",
                        "nullable": true
                    },
                    "mobile": {
                        "type": "string",
                        "example": "+1234567890",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    },
                    "email_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "super-admin"
                        },
                        "nullable": true
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "users.view"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum_admin": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Admin authentication token (Guard: api_admin). Use login endpoint to get token."
            },
            "sanctum_user": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "User authentication token (Guard: api_user). Use login endpoint to get token."
            },
            "sanctum_tenant_user": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Tenant user authentication token (Guard: api_tenant_user). Use login endpoint to get token."
            }
        }
    },
...
}